src/third_party/blink/web_tests/http/tests/mojo/lite/imports.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings_lite.js"></script>
<script src="/gen/mojo/public/js/ts/bindings/tests/export1.test-mojom-lite.js"></script>
<script src="/gen/mojo/public/js/ts/bindings/tests/export2.test-mojom-lite.js"></script>
<script src="/gen/mojo/public/js/ts/bindings/tests/export3.test-mojom-lite.js"></script>
<script src="/gen/mojo/public/js/ts/bindings/tests/export4.test-mojom-lite.js"></script>
<script src="/gen/mojo/public/js/ts/bindings/tests/import.test-mojom-lite.js"></script>
<script>
'use strict';
test(() => {
  assert_equals(mojo.tstest.imported.IMPORTED_ONE, 1);
  assert_equals(mojo.tstest.imported.IMPORTED_TWO, 2);
  assert_equals(mojo.tstest.imported.IMPORTED_THREE, 3);
}, 'Checks that values of imported constants is correct');
</script>
